+++ packages/contracts/contracts/Interfaces/IBorrowerOperations.sol
@@ -14,41 +14,41 @@ interface IBorrowerOperations {
     event EBTCTokenAddressChanged(address _ebtcTokenAddress);
     event FeeRecipientAddressChanged(address _feeRecipientAddress);
     event CollateralAddressChanged(address _collTokenAddress);
-    event FlashLoanSuccess(address _receiver, address _token, uint _amount, uint _fee);
+    event FlashLoanSuccess(address _receiver, address _token, uint256 _amount, uint256 _fee);
 
     // --- Functions ---
 
     function openCdp(
-        uint _EBTCAmount,
+        uint256 _EBTCAmount,
         bytes32 _upperHint,
         bytes32 _lowerHint,
-        uint _stEthBalance
+        uint256 _stEthBalance
     ) external returns (bytes32);
 
     function addColl(
         bytes32 _cdpId,
         bytes32 _upperHint,
         bytes32 _lowerHint,
-        uint _stEthBalanceIncrease
+        uint256 _stEthBalanceIncrease
     ) external;
 
     function withdrawColl(
         bytes32 _cdpId,
-        uint _stEthBalanceDecrease,
+        uint256 _stEthBalanceDecrease,
         bytes32 _upperHint,
         bytes32 _lowerHint
     ) external;
 
     function withdrawEBTC(
         bytes32 _cdpId,
-        uint _amount,
+        uint256 _amount,
         bytes32 _upperHint,
         bytes32 _lowerHint
     ) external;
 
     function repayEBTC(
         bytes32 _cdpId,
-        uint _amount,
+        uint256 _amount,
         bytes32 _upperHint,
         bytes32 _lowerHint
     ) external;
@@ -57,8 +57,8 @@ interface IBorrowerOperations {
 
     function adjustCdp(
         bytes32 _cdpId,
-        uint _stEthBalanceDecrease,
-        uint _debtChange,
+        uint256 _stEthBalanceDecrease,
+        uint256 _debtChange,
         bool isDebtIncrease,
         bytes32 _upperHint,
         bytes32 _lowerHint
@@ -66,12 +66,12 @@ interface IBorrowerOperations {
 
     function adjustCdpWithColl(
         bytes32 _cdpId,
-        uint _stEthBalanceDecrease,
-        uint _debtChange,
+        uint256 _stEthBalanceDecrease,
+        uint256 _debtChange,
         bool isDebtIncrease,
         bytes32 _upperHint,
         bytes32 _lowerHint,
-        uint _stEthBalanceIncrease
+        uint256 _stEthBalanceIncrease
     ) external;
 
     function claimSurplusCollShares() external;
