+++ packages/contracts/contracts/Interfaces/ICollSurplusPool.sol
@@ -10,20 +10,20 @@ interface ICollSurplusPool {
     event ActivePoolAddressChanged(address _newActivePoolAddress);
     event CollateralAddressChanged(address _collTokenAddress);
 
-    event SurplusCollSharesUpdated(address indexed _account, uint _newBalance);
-    event CollSharesTransferred(address _to, uint _amount);
+    event SurplusCollSharesUpdated(address indexed _account, uint256 _newBalance);
+    event CollSharesTransferred(address _to, uint256 _amount);
 
-    event SweepTokenSuccess(address indexed _token, uint _amount, address indexed _recipient);
+    event SweepTokenSuccess(address indexed _token, uint256 _amount, address indexed _recipient);
 
     // --- Contract setters ---
 
-    function getTotalSurplusCollShares() external view returns (uint);
+    function getTotalSurplusCollShares() external view returns (uint256);
 
-    function getSurplusCollShares(address _account) external view returns (uint);
+    function getSurplusCollShares(address _account) external view returns (uint256);
 
-    function increaseSurplusCollShares(address _account, uint _amount) external;
+    function increaseSurplusCollShares(address _account, uint256 _amount) external;
 
     function claimSurplusCollShares(address _account) external;
 
-    function increaseTotalSurplusCollShares(uint _value) external;
+    function increaseTotalSurplusCollShares(uint256 _value) external;
 }
