+++ packages/contracts/contracts/Interfaces/ISortedCdps.sol
@@ -9,7 +9,7 @@ interface ISortedCdps {
     event CdpManagerAddressChanged(address _cdpManagerAddress);
     event SortedCdpsAddressChanged(address _sortedDoublyLLAddress);
     event BorrowerOperationsAddressChanged(address _borrowerOperationsAddress);
-    event NodeAdded(bytes32 _id, uint _NICR);
+    event NodeAdded(bytes32 _id, uint256 _NICR);
     event NodeRemoved(bytes32 _id);
 
     // --- Functions ---
